Fix bug of some transform ops do not return float16 data if out_dtype is set to float16 (#240)
*Issue #, if available:*
#239 

*Description of changes:*


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

+++ python/graphstorm/gconstruct/transform.py
@@ -442,7 +442,8 @@ def parse_feat_ops(confs):
                 transform = NumericalMinMaxTransform(feat['feature_col'],
                                                      feat_name,
                                                      max_bound,
-                                                     min_bound)
+                                                     min_bound,
+                                                     out_dtype=out_dtype)
             else:
                 raise ValueError('Unknown operation: {}'.format(conf['name']))
         ops.append(transform)
